+++ Project.toml
@@ -13,12 +13,14 @@ Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 
 [weakdeps]
+AbstractFFTs = "621f4979-c628-5d54-868e-fcf4e3e8185c"
 ChainRulesCore = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"
 CovarianceEstimation = "587fd27a-f159-11e8-2dae-1979310e6154"
 Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 Tracker = "9f7883ad-71c0-57eb-9f7f-b5c9e6d3789c"
 
 [extensions]
+AbstractFFTsExt = "AbstractFFTs"
 ChainRulesCoreExt = "ChainRulesCore"
 CovarianceEstimationExt = "CovarianceEstimation"
 TrackerExt = "Tracker"
